# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeGate",
    "description": "Generative AI CodeGen security gateway",
    "version": "0.1.7"
  },
  "paths": {
...
    "/api/v1/workspaces": {
      "get": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "List Workspaces",
        "description": "List all workspaces.",
        "operationId": "v1_list_workspaces",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListWorkspacesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Create Workspace",
        "description": "Create a new workspace.",
        "operationId": "v1_create_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/active": {
      "get": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "List Active Workspaces",
        "description": "List all active workspaces.\n\nIn it's current form, this function will only return one workspace. That is,\nthe globally active workspace.",
        "operationId": "v1_list_active_workspaces",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListActiveWorkspacesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Activate Workspace",
        "description": "Activate a workspace by name.",
        "operationId": "v1_activate_workspace",
        "parameters": [
          {
            "name": "status_code",
            "in": "query",
            "required": false,
            "schema": {
              "default": 204,
              "title": "Status Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_name}": {
      "delete": {
        "tags": [
          "CodeGate API",
          "Workspaces"
        ],
        "summary": "Delete Workspace",
        "description": "Delete a workspace by name.",
        "operationId": "v1_delete_workspace",
        "parameters": [
          {
            "name": "workspace_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActivateWorkspaceRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ActivateWorkspaceRequest"
      },
      "ActiveWorkspace": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_updated": {
            "title": "Last Updated"
          }
        },
        "type": "object",
        "required": [
          "name",
          "is_active",
          "last_updated"
        ],
        "title": "ActiveWorkspace"
      },
...
      "CreateWorkspaceRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateWorkspaceRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListActiveWorkspacesResponse": {
        "properties": {
          "workspaces": {
            "items": {
              "$ref": "#/components/schemas/ActiveWorkspace"
            },
            "type": "array",
            "title": "Workspaces"
          }
        },
        "type": "object",
        "required": [
          "workspaces"
        ],
        "title": "ListActiveWorkspacesResponse"
      },
      "ListWorkspacesResponse": {
        "properties": {
          "workspaces": {
            "items": {
              "$ref": "#/components/schemas/Workspace"
            },
            "type": "array",
            "title": "Workspaces"
          }
        },
        "type": "object",
        "required": [
          "workspaces"
        ],
        "title": "ListWorkspacesResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Workspace": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "is_active"
        ],
        "title": "Workspace"
      }
    }
  }
}